Broke the contract again. Adding features: you can now talk to the bot privately.
